Move package to comply with new plugin model
