+++ js/src/main/kotlin/org/rust/js/toml/schema/RsCargoManifestSchemaProvider.kt
@@ -3,21 +3,21 @@
  * found in the LICENSE file.
  */
 
-package org.rust.toml.schema
+package org.rust.js.toml.schema
 
 import com.intellij.openapi.vfs.VirtualFile
 import com.jetbrains.jsonSchema.extension.JsonSchemaFileProvider
 import com.jetbrains.jsonSchema.extension.JsonSchemaProviderFactory
 import com.jetbrains.jsonSchema.extension.SchemaType
 
-class RsCargoConfigSchemaProvider : JsonSchemaFileProvider {
+class RsCargoManifestSchemaProvider : JsonSchemaFileProvider {
     override fun isAvailable(file: VirtualFile): Boolean =
-        file.path.endsWith(".cargo/config.toml")
+        file.name == "Cargo.toml"
 
-    override fun getName(): String = "Cargo Config"
+    override fun getName(): String = "Cargo Manifest"
 
     override fun getSchemaFile(): VirtualFile? =
-        JsonSchemaProviderFactory.getResourceFile(javaClass, "/jsonSchemas/cargo-config-schema.json")
+        JsonSchemaProviderFactory.getResourceFile(this::class.java, "/jsonSchemas/cargo-manifest-schema.json")
 
     override fun getSchemaType(): SchemaType = SchemaType.embeddedSchema
 }
